Commit before merging master
